created test game with enemies but they can't hurt you, you win if you kill all enemies

+++ Game.cpp
@@ -3,7 +3,9 @@
 
 Game::Game() : player(45,24) , field(90,28)
 {
-	generateWalls(12);
+	points = 0;
+	generateWalls();
+	generateTanks();
 	drawWalls();
 	
 }
@@ -11,12 +13,15 @@ Game::~Game()
 {
 }
 void Game::gameLoop() {
-	while (true) {
+	while (enemyTanks.size() != 0) {
 		field.drawObject(player);
-		field.drawObject(playerShell);
+		
 		field.print();
+		printGameStatus();
 
-		std::this_thread::sleep_for(std::chrono::microseconds(100000));
+		std::this_thread::sleep_for(std::chrono::microseconds(10000));
+		processHits();
+		processTanks();
 		field.deleteObject(player);
 		if (_kbhit()) {
 			int button = static_cast<int>(_getch());
@@ -28,14 +33,47 @@ void Game::gameLoop() {
 				player.control(button);
 				if (field.isBarrier(player) == false)
 					player.move();
-			}	
-		}
-		field.deleteObject(playerShell);
-		if (field.isBarrier(playerShell) == false)
-			playerShell.move();
-		else
+			}
+		}			
+	}
+	system("cls");
+	std::cout << "YOU WON" << std::endl;
+	std::cout << "POINTS : " << std::to_string(points);
+}
+void Game::processHits() {
+	auto i = enemyTanks.begin();
+	while ( i != enemyTanks.end() )
+		if (i->isWounded(playerShell) == true) {
+			points++;
+			field.deleteObject(*i);
+			i = enemyTanks.erase(i);		
+			field.deleteObject(playerShell);
 			playerShell = Shell(0, 0, 0, 0);
-					
+			break;
+		}
+		else {
+			i++;
+		}
+	field.deleteObject(player);
+	if (field.isBarrier(playerShell) == true) {
+		field.deleteObject(playerShell);
+		playerShell = Shell(0, 0, 0, 0);
+	}
+	field.drawObject(player);
+	field.deleteObject(playerShell);
+	playerShell.move();
+	field.drawObject(playerShell);	
+}
+void Game::processTanks() {
+	for (auto i = enemyTanks.begin(); i != enemyTanks.end(); i++) {
+		if (field.isBarrier(*i) == true) {
+			i->rotate();
+		}		
+		else {
+			field.deleteObject(*i);
+			i->move();
+			field.drawObject(*i);
+		}			
 	}
 }
 void Game::drawWalls() {
@@ -62,11 +100,25 @@ void Game::deleteShells() {
 		field.deleteObject(*i);
 	field.deleteObject(playerShell);
 }
-void Game::generateWalls(int n) {
+void Game::generateTanks() {
+	enemyTanks.push_back(EnemyTank(2,2,1,0));
+	enemyTanks.push_back(EnemyTank(2, 9, 0, 1));
+	enemyTanks.push_back(EnemyTank(84, 1, 0, 1));
+	enemyTanks.push_back(EnemyTank(45, 1, 1, 0));
+	enemyTanks.push_back(EnemyTank(84, 17, -1, 0));
+}
+void Game::generateWalls() {
 	walls.push_back(Wall(0, 7, 20, Wall::GORIZONTAL));
 	walls.push_back(Wall(44, 0, 7, Wall::VERTICAL));
 	walls.push_back(Wall(12, 21, 12, Wall::GORIZONTAL));
 	walls.push_back(Wall(75, 20, 15, Wall::GORIZONTAL));
 	walls.push_back(Wall(50, 15, 18, Wall::GORIZONTAL));
 	walls.push_back(Wall(70, 4, 7, Wall::VERTICAL));
+}
+void Game::printGameStatus() {
+	std::string status = "LIVE : ";
+	status += std::to_string(player.getHelth());
+	status += "\tPOINTS : ";
+	status += std::to_string(points);
+	std::cout << status;
 }
